+++ src/drv_ak8975.h
@@ -2,4 +2,4 @@
 
 bool ak8975detect(sensor_t *mag);
 void ak8975Init(sensor_align_e align);
-void ak8975Read(int16_t *magData);
+void ak8975Read(int32_t *magData);
